Module Rdf

module Bnode : sig ... end

Blank nodes are anonymous labelled nodes in a graph.

module Graph : sig ... end

Graphs are sets of triples.

module Iri : sig ... end

Internationalized Resource Identifiers (IRI) are Unicode strings that identify resources in graphs.

module Literal : sig ... end

Literals are used in graphs for values such as strings, numbers or dates.

module Triple : sig ... end

Triples are sets of entities that codifies statements about resources.

module Xsd : sig ... end

XML Schema Definition (XSD) contains standard datatypes for literals.