Module Rdf.Bnode

Blank nodes are anonymous labelled nodes in a graph.

see https://www.w3.org/TR/rdf11-concepts/#section-blank-nodes

type internal

type t = [

| `Bnode of internal

]

Initialization

val create : unit -> [> t ]

Creates a new blank node with a random label.

The label is constructed from a v4 (random) UUID.

see https://www.itu.int/ITU-T/recommendations/rec.aspx?rec=11746

Comparison

val equal : t -> t -> bool

val compare : t -> t -> int

String operations

val of_string : string -> [> t ]

val to_string : t -> string